GoldTier/requirements.txt
groq>=0.9.0
python-dotenv>=1.0.0
requests>=2.31.0
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.0
httpx>=0.27.0
# Odoo integration (JSON-RPC, no extra library needed – uses requests)
# Social media clients
tweepy>=4.14.0
facebook-sdk>=3.1.0
